+++ botshot/bootstrap/bots/default.yaml.tpl
