--- botshot/bootstrap/urls.py-tpl
@@ -1,8 +1,0 @@
-from django.contrib import admin
-from django.urls import path, include
-
-urlpatterns = [
-    path('admin/', admin.site.urls),
-    path('bots/', include('botshot.urls')),
-    path('', include('botshot.webgui.urls')),
-]
