+++ botshot/bootstrap/urls.py.tpl
@@ -1,0 +1,8 @@
+from django.urls import path, include
+from . import views
+
+urlpatterns = [
+    path('botshot/', include('botshot.urls')),
+    path('chat/', include('botshot.webgui.urls')),
+    path('', views.landing),
+]
